+++ src/components/index.js
@@ -1,3 +1,4 @@
 export * from "./Player";
 export { default as HostPlayer } from "./HostPlayer";
 export { default as ListenerPlayer } from "./ListenerPlayer";
+export { default as Player } from "./Player";
